+++ _source/alias/media.sh
@@ -5,3 +5,8 @@ if [[ -f /usr/bin/cyanrip  ]];
 then
   alias ripCdAsMp3=net_bazzline_media_rip_cd_as_mp3
 fi
+
+if [[ -f /usr/bin/vobcopy ]];
+then
+  alias ripDvdAsVob=vobcopy -M
+fi